Created PHP files and Website Plan
We ended up not deciding to do PHP because it was difficult to find a server that would host some PHP functions that we wished to test. Therefore, we decided to create the bare skeleton of the website and ask George from Kaidoora to do the more detailed work. We also wrote a long plan ahead for our website.